pyproject.toml
[project]
name = "nexus-writer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aerich>=0.8.1",
    "apscheduler>=3.11.0",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "boto3>=1.39.14",
    "bs4>=0.0.2",
    "cryptography>=45.0.3",
    "duckdb>=1.3.2",
    "email-validator>=2.2.0",
    "fastapi>=0.135.0",
    "langchain[google-genai]>=1.0.3",
    "langchain-community>=0.4.1",
    "langchain-neo4j>=0.6.0",
    "langchain-openai>=1.0.1",
    "loguru>=0.7.3",
    "numpy>=2.2.6",
    "pandas>=2.3.1",
    "passlib>=1.7.4",
    "arq>=0.26",
    "pydantic-settings>=2.9.1",
    "python-socketio>=5.13.0",
    "redis[hiredis]>=4.2.0,<6",
    "socketio>=0.2.1",
    "tortoise-orm[asyncpg]>=0.22.0",
    "uvicorn>=0.34.3",
    "toonify>=1.5.1",
    "pymongo>=4.16.0",
    "pyyaml>=6.0",
    "langgraph>=1.0.4",
    "langchain-anthropic>=1.3.3",
    "tenacity>=9.1.2",
]

[tool.aerich]
tortoise_orm = "src.infrastructure.db.postgres.TORTOISE_ORM"
location = "./migrations"
src_folder = "."
